feedviz/requirements.txt
crewai==1.14.1
crewai-tools==1.14.1
openai==1.54.3
langchain-openai==0.2.3
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pandas==2.2.3
numpy==2.1.3
openpyxl==3.1.5
nltk==3.9.1
spacy==3.8.2
vaderSentiment==3.3.2
plotly==5.24.1
streamlit==1.40.1
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
gspread==6.1.4
google-auth==2.36.0